优化老接口,且增加相册管理接口

+++ data/response.go
@@ -3,6 +3,6 @@ package data
 //正常search返回结果
 type Response struct {
 	Status int         `json:"status"`
-	Cost   int64       `json:"cost"`
+	Cost   int         `json:"cost"`
 	Data   interface{} `json:"data"` //保存返回数据
 }
